+++ packages/terra-core-docs/src/terra-dev-site/doc/demographics-banner/example/DemographicsBannerDeceased.jsx
@@ -8,7 +8,7 @@ const DemographicsBannerDeceased = () => (
     age="25 Years"
     dateOfBirth="May 9, 1992"
     deceasedDate="March 12, 2017"
-    gender="True Male"
+    gender="Male"
     identifiers={{ MRN: 12343, REA: '3JSDA' }}
     photo={<Image alt="My Cat" src={demographicsImage} />}
     personName="Johnathon Doe"