+++ packages/terra-outline-view-table/src/index.js
@@ -1,0 +1,3 @@
+import OutlineViewTable from './OutlineViewTable';
+
+export default OutlineViewTable;