+++ packages/terra-action-header/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 2.75.0 - (December 7, 2022)
+
+* Changed
+  * Minor dependency version bump
+
 ## 2.74.3 - (July 5, 2022)
 
 * Changed