+++ packages/terra-arrange/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 3.50.0 - (December 7, 2022)
+
+* Changed
+  * Minor dependency version bump
+
 ## 3.49.0 - (September 28, 2021)
 
 * Changed