+++ packages/terra-card/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 3.44.0 - (December 7, 2022)
+
+* Changed
+  * Minor dependency version bump
+
 ## 3.43.0 - (February 2, 2022)
 
 * Changed