+++ packages/terra-dynamic-grid/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 3.34.0 - (December 7, 2022)
+
+* Changed
+  * Minor dependency version bump
+
 ## 3.33.1 - (December 14, 2021)
 
 * Changed