+++ packages/terra-grid/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 6.33.0 - (December 7, 2022)
+
+* Changed
+  * Minor dependency version bump
+
 ## 6.32.1 - (December 14, 2021)
 
 * Changed