+++ packages/terra-html-table/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 1.15.0 - (December 7, 2022)
+
 * Changed
   * Added ability to add row headers to the table.