+++ packages/terra-hyperlink/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 2.54.0 - (December 7, 2022)
+
+* Changed
+  * Minor dependency version bump
+
 ## 2.53.0 - (July 14, 2022)
 
 * Fixed