+++ packages/terra-i18n/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 4.39.0 - (December 7, 2022)
+
+* Changed
+  * Minor dependency version bump
+
 ## 4.38.1 - (January 4, 2022)
 
 * Changed