+++ packages/terra-icon/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 3.49.0 - (December 7, 2022)
+
 * Changed
   * Updated to `one-cerner-style-icons` v1.49.0