+++ packages/terra-list/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## Unreleased
 
+## 4.53.0 - (January 3, 2023)
+* Changed
+  * Added a `ariaSelectionStyle` prop that provides accessibility support for single and multi select lists
+  * Fix section and subsection header colors in the lowlight theme.
+
+* Added
+  * Prop `ariaDescribedBy` - Associates a list with an element or elements that convey information about the list and how to interact with it.
+  * Prop `ariaDescription` - Conveys information about a list and how to interact with it.
+  * Prop `ariaDetails` - Associates a list with an element or elements that convey information about the list and how to interact with it.
+
+## 4.52.0 - (December 7, 2022)
+
 * Changed
   * Sections and subsections updated to add programmatic context to communicate hierarchal relationships