+++ packages/terra-section-header/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 2.57.0 - (December 7, 2022)
+
 * Changed
   * Renamed title prop to text to avoid conflict with native html attribute