+++ packages/terra-toggle-section-header/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 2.59.0 - (December 7, 2022)
+
 * Changed
   * Renamed `terra-section-header` title prop to text
   * Updated snapshots from changes made to terra-section-header